requirements.txt
streamlit
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
numpy
beautifulsoup4
certifi